chore(npm-tools): extract DSL for JPSs to a module
This is to be able to reuse it in more than one lexer.
